+++ app/admin/user.rb
@@ -1,6 +1,7 @@
 ActiveAdmin.register User do
-  permit_params :email, :name, :initials, :username, :locale, :time_zone, :authy_enabled,
-                :password, :password_confirmation
+  permit_params :email, :name, :initials, :username, :locale,
+    :time_zone, :authy_enabled, :password, :password_confirmation,
+    :role
 
   index do
     selectable_column