+++ app/controllers/users_controller.rb
@@ -1,5 +1,5 @@
 class UsersController < ApplicationController
-  before_action :set_project
+  before_action :set_project, except: :create
 
   respond_to :html, :json
 
@@ -10,28 +10,24 @@ class UsersController < ApplicationController
   end
 
   def create
-    user_params = allowed_params.merge(was_created: true)
-    @user = User.create_with(user_params).find_or_create_by(email: user_params[:email])
-    @current_team_users = current_team_users
-
+    @user = User.new(
+      email: allowed_params[:email],
+      name: allowed_params[:name],
+      initials: allowed_params[:initials],
+      username: allowed_params[:username]
+    )
     authorize @user
 
-    if @user.persisted? || @user.save
-      if policy_scope(User).include?(@user)
-        flash[:alert] = t 'is already a member of this project', scope: 'users', email: @user.email
-      else
-        policy_scope(User) << @user
-        @user.teams << current_team unless @user.teams.include?(current_team)
-        action = @user.was_created ? 'sent an invite to join' : 'added to'
-        flash[:notice] = t "was #{action} this project", scope: 'users', email: @user.email
-      end
-
-      respond_to do |format|
-        format.js { render :refresh_user_list }
-        format.html { redirect_to project_users_url(@project) }
-      end
+    if @user.save
+      @user.teams << current_team unless @user.teams.include?(current_team)
+      flash[:notice] = I18n.t('was added to the team', scope: 'users', email: @user.email)
     else
-      render 'index'
+      flash[:alert] = I18n.t('was not created', scope: 'users', email: @user.email)
+    end
+
+    respond_to do |format|
+      format.js { render :refresh_user_list }
+      format.html { redirect_to :back }
     end
   end